Refactored library interface to padd by reference
Matrix objects are passed as the first paramater by reference.
The first paramater is always the object that is operated on.
The result of the operation is again stored in the left-hand side.
A = A * B translates to SMOL_Multiply(&A, &B).
Every function returns an integer status code, reporting the type
of error occured.