Append Column/Row, Copy sub-matrix, invert matrix.
Implemented appending columns and rows to a matrix by
copying the elements of the rhs matrix into the extended
lhs matrix. Alternative method for copying a matrix is introduced
by giving the beginning and end rows (including) of the desired sub-matrix
as parameters to SMOL_CopySubMatrix.

Matrices can now be inverted by applied gaussian elimination. There are still
some inconsistencies with this method. Needs to be examined further.